Thread safety disabled due to SIGBUS error
After having struggled to find the cause of the SIGBUS error doing head-checkout I tracked it down to coming from the pthread library. Cloning now works fine after
I disabled threadsafety in the CMAKE options for the libgit2 build process. Not sure what implications this will have but will keep it like this in order to be
able to keep on working with other aspects of the application.